fix(gateway): address purge review findings
Move request-log purge command/runtime code into its own module, keep dry-run migration checks non-mutating, and make the recurring guard UTC-day based.

Delete libsql request-log children explicitly during purge and report matched counts from the pre-delete count for destructive purges.

Update request-log purge documentation and add production purge mise tasks.
